ingesta-3/requirements.txt
boto3==1.40.36
boto3-stubs==1.40.44
botocore==1.40.36
botocore-stubs==1.40.33
jmespath==1.0.1
mypy-boto3-s3==1.40.26
mysql-connector-python==9.4.0
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
types-awscrt==0.27.6
types-s3transfer==0.13.1
urllib3==2.5.0
